+++ sphinx/ext/napoleon/docstring.py
@@ -8,17 +8,19 @@ import inspect
 import re
 from functools import partial
 from itertools import starmap
-from typing import TYPE_CHECKING, Any, Callable
+from typing import TYPE_CHECKING
 
 from sphinx.locale import _, __
 from sphinx.util import logging
-from sphinx.util.typing import get_type_hints, stringify_annotation
+from sphinx.util.typing import RenderMode, get_type_hints, stringify_annotation
 
 if TYPE_CHECKING:
-    from collections.abc import Iterator
+    from collections.abc import Callable, Iterator
+    from typing import Any
 
     from sphinx.application import Sphinx
     from sphinx.config import Config as SphinxConfig
+    from sphinx.util.typing import _StringifyMode
 
 logger = logging.getLogger(__name__)
 
@@ -880,8 +882,10 @@ class GoogleDocstring:
                                    ) or {})
                     self._annotations = get_type_hints(self._obj, None, localns)
                 if _name in self._annotations:
-                    return stringify_annotation(self._annotations[_name],
-                                                'fully-qualified-except-typing')
+                    mode: _StringifyMode = RenderMode.fully_qualified_except_typing
+                    if getattr(self._config, 'python_display_short_literal_types', None):
+                        mode = RenderMode.fully_qualified_except_typing_short_literal
+                    return stringify_annotation(self._annotations[_name], mode)
         # No annotation found
         return ""